+++ objctidelib/SDTideFactory.h
@@ -28,9 +28,9 @@
 
 + (SDTide*)todaysTidesForStationName:(NSString*)name;
 + (NSArray*)tidesForStationName:(NSString*)name;
-+ (NSArray*)tidesForStationName:(NSString*)name withInterval:(int)interval forDays:(int)days;
++ (NSArray*)tidesForStationName:(NSString*)name withInterval:(long)interval forDays:(long)days;
 + (SDTide*)tidesForStationName:(NSString*)name fromDate:(NSDate*)fromDate toDate:(NSDate*)toDate;
-+ (SDTide*)tideForStationName:(NSString*)name withInterval:(int)interval fromDate:(NSDate*)fromDate toDate:(NSDate*)toDate;
++ (SDTide*)tideForStationName:(NSString*)name withInterval:(long)interval fromDate:(NSDate*)fromDate toDate:(NSDate*)toDate;
 + (SDTideStationData*)tideStationWithName:(NSString*)name;
 
 @end
